Subject: Serializes casadi function wrapper.
Detailed Info: As the fabrics planner only needs the
casadiFunctionWrapper at runtime, its serialization only requires the
serialization of that wrapper. This is done in this commit.
Currently, this requires the recreation of the expressions dict, which
might be removed in the future as it is not needed at runtime.

Issue: #8