ft[edf]: Adds explicit differential map for edf.
The integration of Euclidean Signed Distance Fields requires to have
non-symbolic differential maps. Thus, they cannot be differentiated
symbolically. To bypass this need, this commit implementents an
ExplicitDifferentialMap which uses symbolic variables for the jacobian
and the time derivative of the jacobian.  At runtime these values have
to be passed to the planner. A simplified parameterized fabric planner
is implemented in the corresponding example.